Move ISO to RL optimizers package
