+++ docs/apidocs/riemannian-optimizers.md
@@ -7,7 +7,4 @@ emerging_optimizers.riemannian_optimizers
 
 .. automodule:: emerging_optimizers.riemannian_optimizers.normalized_optimizer
     :members:
-
-.. automodule:: emerging_optimizers.riemannian_optimizers.isospectral
-    :members:
 ```
