+++ docs/apidocs/rl-optimizers.md
@@ -1,0 +1,10 @@
+```{eval-rst}
+.. role:: hidden
+    :class: hidden-section
+
+emerging_optimizers.rl_optimizers
+=================================
+
+.. automodule:: emerging_optimizers.rl_optimizers.isospectral
+    :members:
+```
