+++ emerging_optimizers/rl_optimizers/__init__.py
@@ -12,4 +12,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from emerging_optimizers.riemannian_optimizers.normalized_optimizer import *
+from emerging_optimizers.rl_optimizers.isospectral import *
